moved Driver and MapDriver to a different package
