+++ splitwise/src/main/java/com/skmonjurul/splitwise/repo/UserRepo.java
@@ -1,10 +1,7 @@
 package com.skmonjurul.splitwise.repo;
 
-import com.skmonjurul.splitwise.model.Group;
+import com.skmonjurul.driver.Driver;
 import com.skmonjurul.splitwise.model.User;
-
-import java.util.HashMap;
-import java.util.Map;
 
 public class UserRepo {
 //    private static Map<String, User> userMap = new HashMap<>();
